pyproject.toml
[project]
name = "hmcp"
version = "0.0.1"
description = "Healthcare MCP"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Innovaccer" }]
dependencies = [
    "asyncio>=3.4.3",
    "langchain-groq>=0.3.2",
    "mcp-use>=1.2.7",
    "mcp[cli]>=1.6.0",
    "nest-asyncio>=1.6.0",
    "fastmcp>=0.1.0",
    "uvicorn>=0.27.0",
    "starlette>=0.36.0",
    "PyJWT>=2.8.0",
    "python-jose>=3.3.0",
    "cryptography>=41.0.0",
    "nemoguardrails==0.13.0",
    "langchain-openai==0.3.14"
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "black==25.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hmcp"]
